requirements.txt
fastapi
uvicorn
openai
PyGithub
requests
python-dotenv
